Remove obsolete output reference validator
